⏮ revert background pricecheckAll - fail for multiple bot
