+++ src/classes/Commands/Commands.ts
@@ -248,6 +248,8 @@ export default class Commands {
             this.review.offerInfo(steamID, message);
         } else if (command === 'pricecheck' && isAdmin) {
             this.request.pricecheckCommand(steamID, message);
+        } else if (command === 'pricecheckall' && isAdmin) {
+            void this.request.pricecheckAllCommand(steamID);
         } else if (command === 'check' && isAdmin) {
             void this.request.checkCommand(steamID, message);
         } else if (command === 'find' && isAdmin) {
